refactor : using SearchBar component
