+++ src/app/(root)/search/page.tsx
@@ -1,59 +1,19 @@
 "use client";
 
-import searchBread from "../../../../public/image/breads/searchBread.png";
-import Image from "next/image";
-import {
-  useRouter,
-  useSearchParams,
-} from "../../../../node_modules/next/navigation";
-import { useEffect, useState, FormEvent } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
+import { useEffect, useState } from "react";
 import { BakeryList } from "@/components/BakeryList/BakeryList";
 import { searchBakery } from "@/app/api/(supabase)/(bakery)/route";
-
-type Bakery = {
-  bakery_id: string;
-  name: string;
-  image: string;
-  phone: string;
-  address: string;
-};
+import SearchBar from "@/components/commons/SearchBar";
+import { Bakery } from "@/types/bakery";
 
 const SearchPage = () => {
   const router = useRouter();
-  const [keyword, setKeyword] = useState("");
-  const [initialKeyword, setInitialKeyword] = useState("");
   const [noResults, setNoResults] = useState(false);
   const [searchResults, setSearchResults] = useState<Bakery[]>([]);
 
-  const search = () => {
-    if (keyword.trim().length < 2) {
-      alert("검색어를 2글자 이상 적어주세요.");
-      return;
-    }
-    router.push(`/search?keyword=${keyword}`);
-  };
-
-  const handleSearch = (event: FormEvent) => {
-    event.preventDefault();
-    search();
-  };
-
-  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === "Enter") {
-      event.preventDefault();
-      search();
-    }
-  };
-
   const searchParams = useSearchParams();
   const searchedKeyword = searchParams.get("keyword");
-
-  useEffect(() => {
-    if (searchedKeyword) {
-      setKeyword(searchedKeyword);
-      setInitialKeyword(searchedKeyword);
-    }
-  }, [searchedKeyword]);
 
   useEffect(() => {
     const fetchData = async () => {
@@ -71,35 +31,15 @@ const SearchPage = () => {
     fetchData();
   }, [searchedKeyword]);
 
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setKeyword(e.target.value);
+  const handleSearch = (keyword: string) => {
+    router.push(`/search?keyword=${keyword}`);
   };
 
   return (
     <>
       <div className="relative">
         <div className="w-screen flex justify-center items-center h-60">
-          <div className="flex relative w-full max-w-2xl items-center ">
-            <input
-              className="w-full p-4 pl-10 pr-20 rounded-full border border-point text-2xl h-12 flex items-center font-secondary focus:outline-none"
-              type="text"
-              value={keyword}
-              onChange={handleChange}
-              onKeyDown={handleKeyDown}
-              placeholder="어떤 대빵이가 궁금하니?"
-            />
-            <button
-              className="absolute right-4 top-1/2 transform -translate-y-1/2"
-              onClick={handleSearch}
-            >
-              <Image
-                src={searchBread}
-                alt="search bread"
-                width={40}
-                height={40}
-              />
-            </button>
-          </div>
+          <SearchBar initialKeyword={searchedKeyword || ""} onSearch={handleSearch} />
         </div>
         <div className="flex justify-center items-center">
           {noResults ? (
