+++ src/types/bakery.ts
@@ -1,0 +1,7 @@
+export type Bakery = {
+  bakery_id: string;
+  name: string;
+  image: string;
+  phone: string;
+  address: string;
+};
