+++ src/types/search.ts
@@ -1,0 +1,4 @@
+export type SearchBarProps = {
+  initialKeyword?: string;
+  onSearch?: (keyword: string) => void;
+};
